move exits to own files
